# Perform major parser refactoring