--- a/src/Toml/Parser.hs
+++ b/src/Toml/Parser.hs
@@ -8,18 +8,22 @@ module Toml.Parser
        ) where
 
 import Control.DeepSeq (NFData)
-import Data.Text (Text, pack)
+import Data.Text (Text)
 import GHC.Generics (Generic)
 
-import Toml.Parser.TOML (tomlP)
+import Toml.Parser.Item (tomlP)
+import Toml.Parser.Validate (validateItems)
 import Toml.Type (TOML)
 
+import qualified Data.Text as T
 import qualified Toml.Parser.Core as P (errorBundlePretty, parse)
 
 
 -- | Pretty parse exception for parsing toml.
-newtype ParseException = ParseException Text
-    deriving (Show, Eq, Generic, NFData)
+newtype ParseException = ParseException
+    { unParseException :: Text
+    } deriving stock (Show, Generic)
+      deriving newtype (Eq, NFData)
 
 {- | Parses 'Text' as 'TOML' AST object. If you want to convert 'Text' to your
 custom haskell data type, use 'Toml.Bi.Code.decode' or 'Toml.Bi.Code.decodeFile'
@@ -27,5 +31,7 @@ functions.
 -}
 parse :: Text -> Either ParseException TOML
 parse t = case P.parse tomlP "" t of
-    Left err   -> Left $ ParseException $ pack $ P.errorBundlePretty err
-    Right toml -> Right toml
+    Left err    -> Left $ ParseException $ T.pack $ P.errorBundlePretty err
+    Right items -> case validateItems items of
+        Left err   -> Left $ ParseException $ T.pack $ show err
+        Right toml -> Right toml