--- a/stack.yaml
+++ b/stack.yaml
@@ -1,9 +1,7 @@
-resolver: lts-13.25
+resolver: lts-14.1
 
 extra-deps:
-  - hedgehog-1.0
   - parser-combinators-1.2.0
-  - tasty-hedgehog-1.0.0.0
   - toml-parser-0.1.0.0
 
   - htoml-megaparsec-2.1.0.3