--- a/tomland.cabal
+++ b/tomland.cabal
@@ -31,12 +31,12 @@ copyright:           2018-2019 Kowainik
 category:            TOML, Text, Configuration
 build-type:          Simple
 extra-doc-files:     README.md
-                   , CHANGELOG.md
+                     CHANGELOG.md
 extra-source-files:  test/golden/*.golden
                    , test/examples/*.toml
 tested-with:         GHC == 8.2.2
-                   , GHC == 8.4.4
-                   , GHC == 8.6.5
+                     GHC == 8.4.4
+                     GHC == 8.6.5
 
 source-repository head
   type:                git
@@ -57,6 +57,8 @@ common common-options
 
   default-language:    Haskell2010
   default-extensions:  DeriveGeneric
+                       DerivingStrategies
+                       GeneralizedNewtypeDeriving
                        InstanceSigs
                        LambdaCase
                        OverloadedStrings
@@ -78,9 +80,11 @@ library
                          Toml.Generic
                          Toml.Parser
                            Toml.Parser.Core
+                           Toml.Parser.Item
+                           Toml.Parser.Key
                            Toml.Parser.String
+                           Toml.Parser.Validate
                            Toml.Parser.Value
-                           Toml.Parser.TOML
                          Toml.PrefixTree
                          Toml.Printer
                          Toml.Type