Resolves part of #5
An alternative solution to #212

This PR completely rewrites the parser internals of `tomland` to an absolutely new scheme (details below). There should be no user-visible API changes since only internals are changed.

I don't expect you to have a complete understanding of the code during the review (because I'm not sure I understand everything completely by myself). However, all tests are passing, I didn't change any tests except broken ones. So, please check for general stuff, like correct documentation, sane module, data types and function names, code style, etc.

**Highlights:**

1. Parser itself became much simpler. `megaparsec` part of `tomland` is now much smaller and simpler.
2. Parser became faster (benchmarks below).
3. Parser became more correct (some bugs are fixed).
4. Parser now performs some validation of TOML (though, tests are not added, I'll leave to a separate issue).
5. Faster compilation (more smaller modules instead of a few big ones).

**New algorithm**

Before:

* `parse :: Text -> TOML`

Now:

* `parse :: Text -> [TomlItem]`
* `validate :: [TomlItem] -> TOML`

`TomlItem` is simply one line in the TOML file (well, almost, if we forget about multiline strings). It represents one independent item. The parser now splits TOML text into items and then grouping algorithm converts this list into a proper tree, and only then this tree is converted to `TOML`. It sounds difficult but this was already done before, however, implicitly by parser, which made parser more complex. Now the parser is much simpler and all hard work is performed by separate pure functions.

Benchmarks:

```
Decode/tomland
time                 277.6 μs   (253.1 μs .. 294.2 μs)
                     0.918 R²   (0.812 R² .. 0.970 R²)
mean                 356.7 μs   (332.7 μs .. 393.9 μs)
std dev              96.83 μs   (63.04 μs .. 134.6 μs)
variance introduced by outliers: 92% (severely inflated)

Decode/htoml
time                 927.6 μs   (906.3 μs .. 974.0 μs)
                     0.991 R²   (0.979 R² .. 0.998 R²)
mean                 910.1 μs   (899.6 μs .. 926.8 μs)
std dev              42.14 μs   (25.18 μs .. 65.44 μs)
variance introduced by outliers: 27% (moderately inflated)

Decode/htoml-megaparsec
time                 422.5 μs   (417.2 μs .. 427.7 μs)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 428.7 μs   (423.9 μs .. 441.4 μs)
std dev              28.52 μs   (7.130 μs .. 54.00 μs)
variance introduced by outliers: 42% (moderately inflated)

Decode/toml-parser
time                 165.3 μs   (163.3 μs .. 167.6 μs)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 166.8 μs   (165.6 μs .. 170.1 μs)
std dev              6.269 μs   (3.502 μs .. 11.16 μs)
variance introduced by outliers: 19% (moderately inflated)
```